NEW: Move ArrayData from View.
Like ViewableData, this is a common dependency of several modules and
so is rightly part of core. In time, we may deprecated it entirely in
favour of allowing other parts of the system (notably the template
engine) to work directly with arrays, but until that time, it belongs
in core.